Move export to PDF handler implementation
-  No need for separate directory for a simple command handler
-  initializeDashboardHandler is separated because it has a good amount of files related

RELATED: RAIL-3492
